Modularize legacy greeting helpers
